+++ test/io/test_cli.py
@@ -261,7 +261,9 @@ def test_invalid_openapi_mode(invalidopenapimodes, defaultenv):
 def test_schema_cache_snapshot(baseenv, key, snapshot_yaml):
     "Dump of schema cache should match snapshot."
 
-    schema_cache = yaml.load(cli(["--dump-schema"], env=baseenv), Loader=yaml.Loader)
+    schema_cache = yaml.load(
+        cli(["--dump-schema-cache"], env=baseenv), Loader=yaml.Loader
+    )
     formatted = yaml.dump(
         schema_cache[key],
         encoding="utf8",